Exercicio separador
Inclusao stack na pasta para otimizacao
