+++ Estrutura_de_Dados/Questionario3/popCaracteres.py
@@ -1,23 +1,4 @@
-class Stack:
-    def __init__(self):
-        self.items = []
-
-    def isEmpty(self):
-        return self.items == [] 
-        
-    def push(self,item):
-        self.items.append(item)
-
-    def pop(self):
-        return self.items.pop()
-    
-    def peek(self):
-        return self.items[-1]
-
-    def size(self):
-        return len(self.items)
-
-
+from stack import Stack
 
 pilha = Stack()
 caracteres = input()
@@ -27,7 +8,7 @@ for e in caracteres:
         saida = saida + str(pilha.pop())
     else:
         pilha.push(e)
-        
+
 print(saida)
