+++ Estrutura_de_Dados/Questionario3/stack.py
@@ -1,0 +1,21 @@
+class Stack:
+    def __init__(self):
+        self.items = []
+
+    def isEmpty(self):
+        return self.items == [] 
+        
+    def push(self,item):
+        self.items.append(item)
+
+    def pop(self):
+        return self.items.pop()
+    
+    def peek(self):
+        return self.items[-1]
+
+    def size(self):
+        return len(self.items)
+
+    
+
